# api/openapi/swagger.yaml
basePath: /api
definitions:
...
  handler.PlanListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/handler.PlanResponse'
        type: array
      total:
        type: integer
    type: object
  handler.PlanResponse:
    properties:
      created_at:
        type: integer
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: integer
    type: object
...
  handler.RoleListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/handler.RoleResponse'
        type: array
      total:
        type: integer
    type: object
  handler.RoleResponse:
    properties:
      created_at:
        type: integer
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: integer
    type: object
...
  response.errorResponse:
    properties:
      code:
        type: integer
      details:
        additionalProperties: true
        type: object
      message:
        example: Internal server error
        type: string
    type: object
  response.invalidParamsResponse:
    properties:
      code:
        type: integer
      details:
        additionalProperties: true
        type: object
      message:
        example: invalid params
        type: string
    type: object
  response.successResponse:
    properties:
      code:
        example: 2000
        type: integer
      data: {}
      message:
        example: 请求成功
        type: string
    type: object
  saas_internal_plan_handler.CreateRequest:
    properties:
      description:
        maxLength: 60
        type: string
      title:
        maxLength: 30
        type: string
    required:
    - title
    type: object
  saas_internal_plan_handler.UpdateRequest:
    properties:
      description:
        maxLength: 60
        type: string
      title:
        maxLength: 30
        type: string
    required:
    - title
    type: object
  saas_internal_role_handler.CreateRequest:
    properties:
      description:
        maxLength: 60
        type: string
      title:
        maxLength: 30
        type: string
    required:
    - title
    type: object
  saas_internal_role_handler.UpdateRequest:
    properties:
      description:
        maxLength: 60
        type: string
      title:
        maxLength: 30
        type: string
    required:
    - title
    type: object
...
host: localhost:8080
info:
  contact:
    email: lirous@lirous.com
    name: Lirous
    url: https://Lirous.com
  description: 自定义描述
  license:
    name: MIT
    url: https://github.com/Lirous587/go-scaffold/main/LICENSE
  termsOfService: https://Lirous.com
  title: 自定义title
  version: "1.0"
paths:
...
  /v1/plan:
    get:
      consumes:
      - application/json
      description: 根据查询参数获取Plan列表，返回当前页数据和total数量
      parameters:
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Plan列表
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.PlanListResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      summary: 获取 Plan 列表
      tags:
      - plan
    post:
      consumes:
      - application/json
      description: 创建新的 Plan
      parameters:
      - description: 创建 Plan 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/saas_internal_plan_handler.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功创建 Plan
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.PlanResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 创建 Plan
      tags:
      - plan
  /v1/plan/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除 Plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除 Plan
          schema:
            $ref: '#/definitions/response.successResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 删除 Plan
      tags:
      - plan
    get:
      consumes:
      - application/json
      description: 读取单条 Plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功创建 Plan
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.PlanResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 读取单条 Plan
      tags:
      - plan
    put:
      consumes:
      - application/json
      description: 根据ID更新 Plan 信息
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新 Plan 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/saas_internal_plan_handler.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新 Plan
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.PlanResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 更新 Plan
      tags:
      - plan
  /v1/role:
    get:
      consumes:
      - application/json
      description: 根据查询参数获取Role列表，返回当前页数据和total数量
      parameters:
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Role列表
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RoleListResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      summary: 获取 Role 列表
      tags:
      - role
    post:
      consumes:
      - application/json
      description: 创建新的 Role
      parameters:
      - description: 创建 Role 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/saas_internal_role_handler.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功创建 Role
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RoleResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 创建 Role
      tags:
      - role
  /v1/role/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除 Role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除 Role
          schema:
            $ref: '#/definitions/response.successResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 删除 Role
      tags:
      - role
    get:
      consumes:
      - application/json
      description: 读取单条 Role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功创建 Role
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RoleResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 读取单条 Role
      tags:
      - role
    put:
      consumes:
      - application/json
      description: 根据ID更新 Role 信息
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新 Role 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/saas_internal_role_handler.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新 Role
          schema:
            allOf:
            - $ref: '#/definitions/response.successResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.RoleResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.invalidParamsResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.errorResponse'
      security:
      - BearerAuth: []
      summary: 更新 Role
      tags:
      - role
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"